Create JpegScanner
the first edition, where I can find files with the extension ozp in the directory (including subdirectories), compare with the copy criteria (while the size is in kilobytes) and copy the appropriate files to the specified directory (which is created if necessary). The root directory, the directory for the copy and file size are set sequentially through the console.